pyproject.toml
[tool.poetry]
name = "reddit_posts"
version = "0.1.0"
description = ""
authors = ["vitorroxo <vitor@mobi2buy.com>"]

[tool.poetry.dependencies]
python = "^3.8"
flask = "1.1.2"
connexion = {version = "^2.7.0", extras = ["swagger-ui"]}
dynaconf = "2.2.3"
requests = "2.24.0"
psycopg2-binary = "2.8.5"
flask-mongoengine = "0.9.5"
flask-apscheduler = "1.11.0"
mongomock = "3.19.0"

[tool.poetry.dev-dependencies]
pytest-mock = "3.1.1"
pytest = "5.4.3"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
